pyproject.toml
[tool.poetry]
name = "synthetic_knowledge_graphs"
version = "0.1.0"
description = ""
authors = ["pablo-sanchez-sony <pablo.sanchez2@sony.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = {version = "^1.26.1", python = ">=3.10,<3.13"}
matplotlib = "^3.8.0"
pyyaml = "^6.0.1"
networkx = "^3.2.1"

[tool.poetry.group.prebuild.dependencies]
flake8 = "^6.1.0"
isort = "^5.12.0"
pytest = "^7.4.3"
mypy = "^1.6.1"


[tool.poetry.group.playground.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#isort
[tool.isort]
line_length = 100
extend_skip = [".md", ".json"]
multi_line_output = 3
